image: change local run/test to mount only src, pom.xml and mvnw
By making this change, developers who have a local user ID other than 1000 can successfully run and test applications locally. This is because, prior to this change the entire `/project/user-app` directory was mounted from localhost. If the local user did not have a 1000 userid, there would be permission errors writing to the mounted host. This change allows the container to own everything except the source folder, pom.xml and mvnw.